# Cache resolved asserters