This tiny patch let us improve performances on atoum tests (this is really clear with inline engine).

_All the metrics above were generated using the [phpunit-extension](https://github.com/atoum/phpunit-extension) and the [PHP-DI test suite](https://github.com/jubianchi/PHP-DI/compare/e6f30a337cdbff13a1dd9ac73c59654ce44177ab...HEAD) (`vendor/bin/atoum -d tests/UnitTest/ tests/IntegrationTest`)_

### Before

```
PHP 7.0.0 (cli) (built: Dec  2 2015 13:06:23) ( NTS )
Copyright (c) 1997-2015 The PHP Group
Zend Engine v3.0.0, Copyright (c) 1998-2015 Zend Technologies
    with blackfire v1.9.0, https://blackfire.io, by Blackfireio Inc.

> Total tests duration: 29.46 seconds.
> Total tests memory usage: 4.00 Mb.
> Running duration: 29.53 seconds.
```

### After

```
PHP 7.0.0 (cli) (built: Dec  2 2015 13:06:23) ( NTS )
Copyright (c) 1997-2015 The PHP Group
Zend Engine v3.0.0, Copyright (c) 1998-2015 Zend Technologies
    with blackfire v1.9.0, https://blackfire.io, by Blackfireio Inc.

> Total tests duration: 1.84 seconds.
> Total tests memory usage: 4.00 Mb.
> Running duration: 1.89 seconds.
```

Here is the diff generated with blackfire (`blackfire run --samples=5 vendor/bin/atoum -d tests/UnitTest/ tests/IntegrationTest`):

![blackfire](https://cloud.githubusercontent.com/assets/327237/12761381/0f6a008a-c9eb-11e5-838b-46152d009fc9.png)