@mageekguy 2016-02-03
And again, a huge warning: the impact of this patch will be less impressive with other engines (fork is main bottleneck with other engines, and with a fork, the cache is not so efficient because it will be recreated in each subprocess)!