@Hywan 2016-02-03
@mageekguy, @jubianchi Maybe we can share the cache between process but it can really slow down the execution due to concurrent I/O. But inside the same test case (so the same process), it will speed up anyway as soon as we use an assertion more than once.