@mageekguy 2016-02-03
@Hywan Cost of serialization/deserialization to transfert cache between each processus will cancel the cache benefit.