@Hywan 2016-02-03
@mageekguy Except if the cache is in a shared memory page: http://php.net/shmop.