@mageekguy 2016-02-03
@Hywan Shared memory page imply specific PHP extension which is rarely available "out of the box" on a workstation and it seems to me that it is unavailable on Windows. Moreover, you must manage lock to use it to avoid race condition. So, this solution add lot of complexity in the code in regards of a small impact on performance and a lack of portability. Not sure that wasting time to implements this kind of tradeoff is a good move, or only for the fun.