@Hywan 2016-02-03
@mageekguy It works on Windows :-).